requirements.txt
langchain==0.0.267
google-generativeai==0.3.1
streamlit==1.24.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2
sentence-transformers==2.2.2
huggingface-hub==0.13.4